requirements.txt
Flask
joblib
pandas
python-dotenv
google-generativeai
